move yoloheadToPredictions() and nms() from evaluate.py to utils.py
